requirements.txt
azure-data-tables
Flask
fastapi
uvicorn
uvloop
httptools
python-multipart
python-jose
cryptography
passlib
bcrypt
azure-cosmos